Cleaned up and moved to the correct package
